+++ Euclid/Sources/Euclid+RealityKit.swift
@@ -5,32 +5,34 @@
 //  Created by Nick Lockwood on 05/10/2022.
 //  Copyright © 2022 Nick Lockwood. All rights reserved.
 //
+//  Distributed under the permissive MIT license
+//  Get the latest version from here:
+//
+//  https://github.com/nicklockwood/Euclid
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//  SOFTWARE.
+//
 
 #if canImport(RealityKit) && swift(>=5.4)
 
+import CoreGraphics
 import RealityKit
-
-#if canImport(UIKit)
-
-import UIKit
-private typealias OSColor = UIColor
-private typealias OSImage = UIImage
-private typealias OSColorComponent = Float
-
-#elseif canImport(AppKit)
-
-import AppKit
-private typealias OSColor = NSColor
-private typealias OSImage = NSImage
-private typealias OSColorComponent = Double
-
-private extension OSImage {
-    var cgImage: CGImage? {
-        cgImage(forProposedRect: nil, context: nil, hints: nil)
-    }
-}
-
-#endif
 
 @available(macOS 10.15, iOS 13.0, *)
 public extension RealityKit.Transform {