+++ Euclid/Sources/Plane.swift
@@ -89,11 +89,11 @@ extension Plane: Codable {
 
 public extension Plane {
     /// A plane located at the origin, aligned with the Y and Z axes.
-    static let yz = Plane(unchecked: Vector(1, 0, 0), w: 0)
+    static let yz = Plane(unchecked: .unitX, w: 0)
     /// A plane located at the origin, aligned with the X and Z axes.
-    static let xz = Plane(unchecked: Vector(0, 1, 0), w: 0)
+    static let xz = Plane(unchecked: .unitY, w: 0)
     /// A plane located at the origin, aligned with the X and Y axes.
-    static let xy = Plane(unchecked: Vector(0, 0, 1), w: 0)
+    static let xy = Plane(unchecked: .unitZ, w: 0)
 
     /// Creates a plane from a point and surface normal.
     /// - Parameters:
@@ -113,7 +113,7 @@ public extension Plane {
     /// > Note: The polygon can be convex or concave. The direction of the plane normal is
     /// based on the assumption that the points are wound in an anti-clockwise direction.
     init?(points: [Vector]) {
-        self.init(points: points, convex: nil)
+        self.init(points: points, convex: nil, closed: nil)
     }
 
     /// Returns the flip-side of the plane.
@@ -201,20 +201,20 @@ internal extension Plane {
         self.init(unchecked: normal, w: normal.dot(pointOnPlane))
     }
 
-    init?(points: [Vector], convex: Bool?) {
+    init?(points: [Vector], convex: Bool?, closed: Bool?) {
         guard !points.isEmpty, !pointsAreDegenerate(points) else {
             return nil
         }
-        self.init(unchecked: points, convex: convex)
+        self.init(unchecked: points, convex: convex, closed: closed)
         // Check all points lie on this plane
         if points.contains(where: { !containsPoint($0) }) {
             return nil
         }
     }
 
-    init(unchecked points: [Vector], convex: Bool?) {
+    init(unchecked points: [Vector], convex: Bool?, closed: Bool?) {
         assert(!pointsAreDegenerate(points))
-        let normal = faceNormalForPolygonPoints(points, convex: convex)
+        let normal = faceNormalForPolygonPoints(points, convex: convex, closed: closed)
         self.init(unchecked: normal, pointOnPlane: points[0])
     }
 
@@ -265,8 +265,10 @@ enum FlatteningPlane: RawRepresentable {
         }
     }
 
-    init(points: [Vector], convex: Bool?) {
-        self.init(normal: faceNormalForPolygonPoints(points, convex: convex))
+    init(points: [Vector]) {
+        self.init(normal: faceNormalForPolygonPoints(
+            points, convex: nil, closed: nil
+        ))
     }
 
     init?(rawValue: Plane) {