+++ ShapeScriptTests/LexerTests.swift
@@ -76,6 +76,25 @@ class LexerTests: XCTestCase {
         }
     }
 
+    func testInsertPhantomParensForDisambiguation() {
+        let input = "-a (b)"
+        let tokens: [TokenType] = [
+            .prefix(.minus), .lparen, .identifier("a"), .rparen,
+            .lparen, .identifier("b"), .rparen, .eof,
+        ]
+        XCTAssertEqual(try tokenize(input).map { $0.type }, tokens)
+    }
+
+    func testInsertPhantomParensForDisambiguation2() {
+        let input = "a + b (c)"
+        let tokens: [TokenType] = [
+            .identifier("a"), .infix(.plus),
+            .lparen, .identifier("b"), .rparen,
+            .lparen, .identifier("c"), .rparen, .eof,
+        ]
+        XCTAssertEqual(try tokenize(input).map { $0.type }, tokens)
+    }
+
     // MARK: numbers
 
     func testZero() {
@@ -520,12 +539,6 @@ class LexerTests: XCTestCase {
         let lc = input.lineAndColumn(at: input.startIndex)
         XCTAssertEqual(lc.line, 1)
         XCTAssertEqual(lc.column, 1)
-        let lc2 = input.lineAndColumn(
-            at: input.startIndex,
-            withLinebreakIndices: input.linebreakIndices
-        )
-        XCTAssertEqual(lc.line, lc2.line)
-        XCTAssertEqual(lc.column, lc2.column)
     }
 
     func testLineAndColumnAtEndOfLine() {
@@ -534,12 +547,6 @@ class LexerTests: XCTestCase {
         let lc = input.lineAndColumn(at: index)
         XCTAssertEqual(lc.line, 2)
         XCTAssertEqual(lc.column, 4)
-        let lc2 = input.lineAndColumn(
-            at: index,
-            withLinebreakIndices: input.linebreakIndices
-        )
-        XCTAssertEqual(lc.line, lc2.line)
-        XCTAssertEqual(lc.column, lc2.column)
     }
 
     func testLineAndColumnAtCRLFEndOfLine() {
@@ -548,11 +555,5 @@ class LexerTests: XCTestCase {
         let lc = input.lineAndColumn(at: index)
         XCTAssertEqual(lc.line, 2)
         XCTAssertEqual(lc.column, 4)
-        let lc2 = input.lineAndColumn(
-            at: index,
-            withLinebreakIndices: input.linebreakIndices
-        )
-        XCTAssertEqual(lc.line, lc2.line)
-        XCTAssertEqual(lc.column, lc2.column)
     }
 }