+++ Viewer/iOS/Document.swift
@@ -49,7 +49,8 @@ class Document: UIDocument {
     }
 
     var errorMessage: NSAttributedString?
-    var accessErrorURL: URL?
+    var errorURL: URL?
+    var isAccessError: Bool = false
     var sourceString: String = ""
     var presets: [Export] = []
 
@@ -85,7 +86,7 @@ class Document: UIDocument {
             forOpeningContentTypes: [.folder],
             asCopy: false
         )
-        picker.directoryURL = accessErrorURL
+        picker.directoryURL = errorURL
         picker.delegate = self
         picker.modalPresentationStyle = .fullScreen
         viewController?.present(picker, animated: true)